+++ app/src/main/java/org/kemea/isafeco/client/streamselector/stubs/output/SessionSourceStreamOutput.java
@@ -3,65 +3,65 @@ package org.kemea.isafeco.client.streamselector.stubs.output;
 import com.google.gson.annotations.SerializedName;
 
 public class SessionSourceStreamOutput {
-	@SerializedName("session_source_service_protocol")
-	String sessionSourceServiceProtocol;
-	@SerializedName("session_source_service_ip")
-	String sessionSourceServiceIp;
-	@SerializedName("session_source_service_port")
-	Integer sessionSourceServicePort;
-	@SerializedName("session_id")
-	Integer sessionId;
-	@SerializedName("session_encryption_key")
-	String sessionEncryptionKey;
-	@SerializedName("session_sdp")
-	String sessionSDP;
+    @SerializedName("session_source_service_protocol")
+    String sessionSourceServiceProtocol;
+    @SerializedName("session_source_service_ip")
+    String sessionSourceServiceIp;
+    @SerializedName("session_source_service_port")
+    Integer sessionSourceServicePort;
+    @SerializedName("session_id")
+    Long sessionId;
+    @SerializedName("session_encryption_key")
+    String sessionEncryptionKey;
+    @SerializedName("session_sdp")
+    String sessionSDP;
 
-	public String getSessionSourceServiceProtocol() {
-		return sessionSourceServiceProtocol;
-	}
+    public String getSessionSourceServiceProtocol() {
+        return sessionSourceServiceProtocol;
+    }
 
-	public void setSessionSourceServiceProtocol(String sessionSourceServiceProtocol) {
-		this.sessionSourceServiceProtocol = sessionSourceServiceProtocol;
-	}
+    public void setSessionSourceServiceProtocol(String sessionSourceServiceProtocol) {
+        this.sessionSourceServiceProtocol = sessionSourceServiceProtocol;
+    }
 
-	public String getSessionSourceServiceIp() {
-		return sessionSourceServiceIp;
-	}
+    public String getSessionSourceServiceIp() {
+        return sessionSourceServiceIp;
+    }
 
-	public void setSessionSourceServiceIp(String sessionSourceServiceIp) {
-		this.sessionSourceServiceIp = sessionSourceServiceIp;
-	}
+    public void setSessionSourceServiceIp(String sessionSourceServiceIp) {
+        this.sessionSourceServiceIp = sessionSourceServiceIp;
+    }
 
-	public Integer getSessionSourceServicePort() {
-		return sessionSourceServicePort;
-	}
+    public Integer getSessionSourceServicePort() {
+        return sessionSourceServicePort;
+    }
 
-	public void setSessionSourceServicePort(Integer sessionSourceServicePort) {
-		this.sessionSourceServicePort = sessionSourceServicePort;
-	}
+    public void setSessionSourceServicePort(Integer sessionSourceServicePort) {
+        this.sessionSourceServicePort = sessionSourceServicePort;
+    }
 
-	public Integer getSessionId() {
-		return sessionId;
-	}
+    public Long getSessionId() {
+        return sessionId;
+    }
 
-	public void setSessionId(Integer sessionId) {
-		this.sessionId = sessionId;
-	}
+    public void setSessionId(Long sessionId) {
+        this.sessionId = sessionId;
+    }
 
-	public String getSessionEncryptionKey() {
-		return sessionEncryptionKey;
-	}
+    public String getSessionEncryptionKey() {
+        return sessionEncryptionKey;
+    }
 
-	public void setSessionEncryptionKey(String sessionEncryptionKey) {
-		this.sessionEncryptionKey = sessionEncryptionKey;
-	}
+    public void setSessionEncryptionKey(String sessionEncryptionKey) {
+        this.sessionEncryptionKey = sessionEncryptionKey;
+    }
 
-	public String getSessionSDP() {
-		return sessionSDP;
-	}
+    public String getSessionSDP() {
+        return sessionSDP;
+    }
 
-	public void setSessionSDP(String sessionSDP) {
-		this.sessionSDP = sessionSDP;
-	}
+    public void setSessionSDP(String sessionSDP) {
+        this.sessionSDP = sessionSDP;
+    }
 
 }
